improve formats of output values
This aims to make the numbers more coherent and formatted for easy readbility.

Also the changes in the CSV number format are so, that the number format is consistent across the entire column, which means that quick imports into typical spreadsheet software will work better, as autodetection of decimal separaters will work.